refactor: migrate PersonaHome for SSR
